+++ src/backend/core/brevo.py
@@ -10,6 +10,47 @@ import requests
 logger = logging.getLogger(__name__)
 
 
+def create_contact_in_brevo(email: str) -> bool:
+    """
+    Create a contact in Brevo.
+
+    Args:
+        email (str): The email address of the user.
+
+    """
+    api_key = settings.BREVO_API_KEY
+    if not api_key:
+        logger.info("Brevo API key not configured: skipping creating contact")
+        return False
+
+    url = "https://api.brevo.com/v3/contacts"
+    headers = {
+        "accept": "application/json",
+        "api-key": api_key,
+        "content-type": "application/json",
+    }
+    payload = {
+        "email": email,
+        "updateEnabled": True,
+    }
+    try:
+        response = requests.post(url, json=payload, headers=headers, timeout=5)
+    except requests.RequestException as e:
+        logger.exception(e)
+        return False
+
+    if not response.ok:
+        logger.error(
+            "Error creating contact in Brevo %s: (%s) %s",
+            email,
+            response.status_code,
+            response.text,
+        )
+        return False
+
+    return True
+
+
 def add_user_to_brevo_list(emails: List[str], list_id: Optional[str]) -> None:
     """
     Add email list to a Brevo list.
@@ -23,6 +64,13 @@ def add_user_to_brevo_list(emails: List[str], list_id: Optional[str]) -> None:
     if not api_key or not list_id:
         logger.info("Brevo API key or list ID not configured: skipping adding contact")
         return
+
+    for email in emails:
+        # Ensure the contact exists before adding to the list
+        # `emails` contains several emails only when used from the admin interface bulk action
+        if not create_contact_in_brevo(email):
+            logger.error("Failed to create contact %s in Brevo, skipping adding to list", email)
+            return
 
     url = f"https://api.brevo.com/v3/contacts/lists/{list_id}/contacts/add"
     headers = {