+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to
 - 🐛(front) code activation fix session end #93
 - 💬(wording) error page wording #102
 - ⚡️(web-search) allow to override returned chunks #107
+- 🐛(activation-codes) create contact in brevo before add to list #108
 
 
 ## [0.0.1] - 2025-10-19